feat: add HomePage component with hero section, features, and footer
